replace landing page with all contributions content * remove old all contributions generator file
