+++ scripts/src/main.js
@@ -22,9 +22,6 @@ const { createCommunityMarkdown } = require('../generators/markdown/community-ma
 
 // Import html generation logic
 const { writeHtmlFiles } = require('../generators/html/quarterly-reports-html-generator');
-const {
-  createAllTimeContributions,
-} = require('../generators/html/all-contributions-html-generator');
 const { createHtmlReports } = require('../generators/html/contributions-report-html-generator');
 const { createIndexHtml } = require('../generators/html/landing-page-html-generator');
 const { createBlogHtml } = require('../generators/html/blog-html-generator');
@@ -62,7 +59,6 @@ async function main() {
   try {
     const commitCacheData = await fs.readFile(commitCacheFile, 'utf8');
     const parsed = JSON.parse(commitCacheData);
-    // parsed expected to be an object mapping prUrlKey -> { firstCommitDate, commitCount } or null
     for (const [k, v] of Object.entries(parsed)) {
       commitCacheFromDisk.set(k, v);
     }
@@ -80,16 +76,10 @@ async function main() {
     console.log('Fetching ongoing review tasks for the Active Workbench...');
     const rawOngoingTasks = await fetchOngoingReviews();
 
-    // Filter out excluded organizations and specific repositories
     const ongoingTasks = rawOngoingTasks.filter((task) => {
       const repoName = task.repo.toLowerCase();
-
-      // Check if it belongs to the open-sauced organization
       const isOpenSauced = repoName.startsWith('open-sauced/');
-
-      // Check for the specific astro-partykit-starter repo
       const isAstroStarter = repoName.includes('astro-partykit-starter');
-
       return !isOpenSauced && !isAstroStarter;
     });
 
@@ -106,7 +96,6 @@ async function main() {
 
     let allContributions = {};
 
-    // Try to load the full contributions data from a JSON file.
     try {
       const data = await fs.readFile(dataFile, 'utf8');
       allContributions = JSON.parse(data);
@@ -119,7 +108,6 @@ async function main() {
       }
     }
 
-    // Determine optimal fetch strategy based on last update time
     const cacheStats = await fs.stat(dataFile).catch(() => null);
     const lastUpdate = cacheStats ? new Date(cacheStats.mtime) : null;
     const today = new Date();
@@ -127,7 +115,6 @@ async function main() {
     let fetchStartYear;
 
     if (!lastUpdate) {
-      // Pass undefined to trigger Auto-Discovery in the fetcher
       fetchStartYear = undefined;
       console.log('First run - triggering auto-discovery of GitHub join date');
     } else {
@@ -149,14 +136,11 @@ async function main() {
         fetchStartYear = currentYear - 1;
         console.log('Last month update - fetching last two years');
       } else {
-        // Fallback for older updates - start from the year before the last update for safety
         fetchStartYear = lastUpdateYear - 1;
         console.log(`Older update - fetching from: ${fetchStartYear}`);
       }
     }
 
-    // If there is no existing contributions data (first run / full fetch), clear
-    // the persistent PR cache so authored PRs are re-processed and repopulated.
     if (!lastUpdate) {
       prCache = new Set();
       console.log(
@@ -164,18 +148,15 @@ async function main() {
       );
     }
 
-    // Merge the persistent commit cache into an in-memory Map and pass it into the fetcher
     const mergedCommitCache = new Map();
     for (const [k, v] of commitCacheFromDisk) mergedCommitCache.set(k, v);
 
-    // Fetch new contributions and update the cache.
     const {
       contributions: newContributions,
       prCache: updatedPrCache,
       commitCache: usedCommitCache,
     } = await fetchContributions(fetchStartYear, prCache, mergedCommitCache);
 
-    // Second pass: merge new contributions and existing ones, enforcing category hierarchy
     let finalContributions = {
       pullRequests: [],
       issues: [],
@@ -184,14 +165,12 @@ async function main() {
       collaborations: [],
     };
 
-    // --- 1. Load Existing Contributions (Preserve categories with category hierarchy) ---
     console.log(
       'Preserving existing contributions by category (enforcing hierarchy: reviewedPrs/coAuthoredPrs > collaborations).'
     );
 
-    const globalLoadedBy = new Map(); // url -> Set of categories
+    const globalLoadedBy = new Map();
 
-    // Load existing data from disk and allow duplication only for reviewedPrs + coAuthoredPrs combo
     const categoryOrder = Object.keys(finalContributions);
     for (const type of categoryOrder) {
       if (Array.isArray(allContributions[type])) {
@@ -208,16 +187,11 @@ async function main() {
           const higherTier = new Set(['reviewedPrs', 'coAuthoredPrs']);
           const currentIsHigher = higherTier.has(type);
 
-          // Allow higher-tier categories to be loaded even if the URL already
-          // exists in `pullRequests` or `issues`. Only prevent loading
-          // `collaborations` when a higher-tier category already exists.
           if (currentIsHigher) {
             finalContributions[type].push(item);
             seen.add(type);
             globalLoadedBy.set(url, seen);
           } else {
-            // Non-higher types (e.g., collaborations) should only be added
-            // when there are no existing higher-tier categories for the URL.
             const hasHigher = Array.from(seen).some((c) => higherTier.has(c));
             if (!hasHigher) {
               finalContributions[type].push(item);
@@ -229,7 +203,6 @@ async function main() {
       }
     }
 
-    // --- 2. Add/Update Newly Fetched Contributions (with hierarchy enforcement) ---
     console.log('Merging newly fetched contributions (enforcing category hierarchy).');
 
     for (const type of Object.keys(newContributions)) {
@@ -257,9 +230,6 @@ async function main() {
           const currentIsHigher = higherTier.has(type);
           const existingInHigher = Array.from(seen).filter((c) => higherTier.has(c));
 
-          // When promoting an item to a higher tier, only remove it from
-          // `collaborations` (the true lower tier). Do NOT remove it from
-          // `pullRequests` or `issues` so merged/authored PRs remain present.
           if (currentIsHigher) {
             const lowerToRemove = ['collaborations'];
             for (const lowerCat of lowerToRemove) {
@@ -270,7 +240,6 @@ async function main() {
             }
 
             finalContributions[type].push(item);
-            // Preserve any existing higher-tier flags and add the current one.
             if (existingInHigher.length > 0) {
               for (const higherCat of existingInHigher) {
                 seen.add(higherCat);
@@ -279,8 +248,6 @@ async function main() {
             seen.add(type);
             globalLoadedBy.set(url, seen);
           } else {
-            // Non-higher types (like collaborations) should only be added if
-            // there is no existing higher-tier representation for the URL.
             if (existingInHigher.length === 0) {
               finalContributions[type].push(item);
               seen.add(type);
@@ -291,48 +258,43 @@ async function main() {
       }
     }
 
-    // --- 3. Sort each category by date ---
     for (const type of Object.keys(finalContributions)) {
       finalContributions[type].sort((a, b) => new Date(b.date) - new Date(a.date));
     }
 
     console.log('Merged and categorized all contributions based on latest status.');
 
-    // Save the updated, full contributions data to a new JSON file
     await fs.writeFile(dataFile, JSON.stringify(finalContributions, null, 2), 'utf8');
     console.log('Updated contributions data saved to file.');
 
     // --- Quarterly grouping, and Markdown and HTML generator functions ---
-    // 1. Group data by quarter
     const grouped = groupContributionsByQuarter(finalContributions);
 
-    // 2. Generate quarterly reports (Markdown)
+    // 1. Generate quarterly reports (Markdown)
     await writeMarkdownFiles(grouped);
 
-    // 3. Generate quarterly reports (HTML)
+    // 2. Generate quarterly reports (HTML)
     const quarterlyHtmlLinks = await writeHtmlFiles(grouped);
 
-    // 4. Generate README and glossary files (Markdown)
+    // 3. Generate README and glossary files (Markdown)
     await createStatsReadme(finalContributions, articles);
 
-    // 5. Generate landing page (index.html)
-    await createIndexHtml();
+    // 4. Generate landing page (index.html)
+    console.log('Generating landing page...');
+    await createIndexHtml(finalContributions, articles);
 
-    // NEW: 5b. Generate the Glossary page
+    // 5. Generate the Glossary page
     console.log('Generating Glossary...');
     await createGlossaryHtml();
 
-    // 6. Generate the Dashboard (All-Time Stats HTML)
-    await createAllTimeContributions(finalContributions, articles);
-
-    // 7. Generate reports page (HTML)
+    // 6. Generate reports page (HTML)
     await createHtmlReports(quarterlyHtmlLinks);
 
-    // 8. Generate Blog Reports (HTML and Markdown)
+    // 7. Generate Blog Reports (HTML and Markdown)
     await createBlogHtml(articles);
     await writeArticlesMarkdown(articles);
 
-    // 9. Generate Community & Activity Reports
+    // 8. Generate Community & Activity Reports
     console.log('Generating Community & Activity reports...');
     await createCommunityHtml(finalContributions, leadershipData, ongoingTasks);
     await createCommunityMarkdown(finalContributions, leadershipData, ongoingTasks);
@@ -341,7 +303,6 @@ async function main() {
     await fs.writeFile(cacheFile, JSON.stringify(Array.from(updatedPrCache)), 'utf8');
     console.log('Updated PR cache saved to file.');
 
-    // Persist the commit cache to disk so future runs reuse it and save API calls.
     try {
       const obj = {};
       for (const [k, v] of usedCommitCache || mergedCommitCache) {
